add test for data reader
