+++ django_project/gap_api/tests/test_measurement_api.py
@@ -5,10 +5,45 @@ Tomorrow Now GAP.
 .. note:: Unit tests for User API.
 """
 
+from datetime import datetime
+from typing import List
+from django.contrib.gis.geos import Point
 from django.urls import reverse
+from unittest.mock import patch
 
 from core.tests.common import FakeResolverMatchV1, BaseAPIViewTest
+from django_project.gap.models import DatasetAttribute
+from django_project.gap.utils.reader import (
+    DatasetReaderValue,
+    DatasetTimelineValue
+)
 from gap_api.api_views.measurement import MeasurementAPI
+from gap.utils.reader import BaseDatasetReader
+from gap.factories import DatasetAttributeFactory
+
+
+class MockDatasetReader(BaseDatasetReader):
+    """Class to mock a dataset reader."""
+
+    def __init__(self, dataset, attributes: List[DatasetAttribute],
+                 point: Point, start_date: datetime,
+                 end_date: datetime) -> None:
+        """Initialize MockDatasetReader class."""
+        super().__init__(dataset, attributes, point, start_date, end_date)
+        self.mocked_data = None
+
+    def get_data_values(self) -> DatasetReaderValue:
+        """Override data values with a mock object."""
+        if self.mocked_data:
+            return self.mocked_data
+        return DatasetReaderValue(
+            {
+                'dataset': [self.dataset.name]
+            },
+            [DatasetTimelineValue(self.start_date, {
+                'test': 100
+            })]
+        )
 
 
 class CommonMeasurementAPITest(BaseAPIViewTest):
@@ -34,7 +69,7 @@ class CommonMeasurementAPITest(BaseAPIViewTest):
         :rtype: WSGIRequest
         """
         request = self.factory.get(
-            reverse('api:v1:user-info') +
+            reverse('api:v1:get-measurement') +
             f'?lat={lat}&lon={lon}&attributes={attributes}'
             f'&start_date={start_dt}&end_date={end_dt}'
         )
@@ -53,6 +88,35 @@ class HistoricalAPITest(CommonMeasurementAPITest):
         response = view(request)
         self.assertEqual(response.status_code, 200)
         self.assertEqual(response.data, {})
+
+    @patch('gap_api.api_views.measurement.get_reader_from_dataset')
+    def test_read_historical_data(self, mocked_reader):
+        """Test read historical data."""
+        view = MeasurementAPI.as_view()
+        mocked_reader.return_value = MockDatasetReader
+        attribute1 = DatasetAttributeFactory.create()
+        attribute2 = DatasetAttributeFactory.create(
+            dataset=attribute1.dataset
+        )
+        attribs = [
+            attribute1.attribute.variable_name,
+            attribute2.attribute.variable_name
+        ]
+        request = self._get_measurement_request(
+            attributes=','.join(attribs)
+        )
+        response = view(request)
+        self.assertEqual(response.status_code, 200)
+        mocked_reader.assert_called_once_with(attribute1.dataset)
+        self.assertIn('metadata', response.data)
+        self.assertIn('data', response.data)
+        response_data = response.data['data']
+        self.assertIn(attribute1.dataset.name, response_data)
+        results = response_data[attribute1.dataset.name]
+        self.assertEqual(len(results), 1)
+        self.assertIn('values', results[0])
+        self.assertIn('test', results[0]['values'])
+        self.assertEqual(100, results[0]['values']['test'])
 
 
 class ForecastAPITest(CommonMeasurementAPITest):
